requirements.txt
lxml==4.4.1
minio==5.0.8
numpy==1.17.3
pdf2image==1.10.0
pdfminer==20191125
pika==1.1.0
Pillow==6.2.1
pycryptodome==3.9.7
pytesseract==0.3.0
reportlab==3.5.32
zipp==2.2.0